[move/#45] 폴더 구조 정리
