+++ shared/feature/my/src/commonMain/kotlin/com/kus/feature/my/state/MyArticleUiState.kt
@@ -1,8 +1,8 @@
-package com.kus.feature.my.ui.state
+package com.kus.feature.my.state
 
 import UiState
 import com.kus.shared.domain.model.my.MyPostItem
 
-data class MyScrapUiState(
+data class MyArticleUiState(
     val articles: UiState<List<MyPostItem>> = UiState.Loading,
 )
