+++ shared/feature/my/src/commonMain/kotlin/com/kus/feature/my/state/MyCommentUiState.kt
@@ -1,4 +1,4 @@
-package com.kus.feature.my.ui.state
+package com.kus.feature.my.state
 
 import UiState
 import com.kus.shared.domain.model.my.MyCommentItem
