+++ shared/feature/my/src/commonMain/kotlin/com/kus/feature/my/type/RestaurantMenu.kt
@@ -1,4 +1,4 @@
-package com.kus.feature.my.ui.type
+package com.kus.feature.my.type
 
 import kustaurant.shared.feature.my.generated.resources.Res
 import kustaurant.shared.feature.my.generated.resources.ic_my_comment
