+++ shared/feature/my/src/commonMain/kotlin/com/kus/feature/my/ui/restaurant/CheckedResViewModel.kt
@@ -4,7 +4,7 @@ import UiError
 import UiState
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.kus.feature.my.ui.state.CheckedUiState
+import com.kus.feature.my.state.CheckedUiState
 import com.kus.shared.domain.my.usecase.GetEvaluatedResUseCase
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
