+++ shared/feature/my/src/commonMain/kotlin/com/kus/feature/my/ui/restaurant/FavoriteResViewModel.kt
@@ -4,7 +4,7 @@ import UiError
 import UiState
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.kus.feature.my.ui.state.FavoriteResUiState
+import com.kus.feature.my.state.FavoriteResUiState
 import com.kus.shared.domain.my.usecase.GetFavoriteResUseCase
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
